Move check_lemmas to prepare_lemma, use that to decide whether or not to augment the lemma data
